Removing references of float and double from BoxedFloat

+++ LanguageKitRuntime/BoxedFloat.h
@@ -3,10 +3,11 @@
 @interface BoxedFloat : NSObject
 {
 	@public
-	double value;
+	// TODO - it would be nice to make it something like NSDecimalNumber
+	NSUInteger value;
 }
 + (BoxedFloat*) boxedFloatWithCString:(const char*) aString;
-+ (BoxedFloat*) boxedFloatWithDouble:(double)aVal;
-+ (BoxedFloat*) boxedFloatWithFloat:(float)aVal;
++ (BoxedFloat*) boxedFloatWithDouble:(NSUInteger)aVal;
++ (BoxedFloat*) boxedFloatWithFloat:(NSUInteger)aVal;
 @end
